pyproject.toml
[project]
name = "mcp-with-sql-agent"
version = "0.1.0"
description = "SQLite assistant built with OpenAI's GPT models and Model Context Protocol (MCP)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "docker>=7.1.0",
    "instructor[openai]>=1.7.3",
    "loguru>=0.7.3",
    "mcp[cli]>=1.3.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "flask>=2.3.0",
    "cx_Oracle>=8.3.0",
    "mysql-connector-python>=8.0.33"
]

[dependency-groups]
dev = ["types-docker>=7.1.0.20241229"]